requirements.txt
azure-iot-device 
gpiozero
bme280
gpiozero-pins
#smbus2 
#RPi.GPIO 
#tkinter